Revert "Generalize Flowdroid XML Parser and MethodAndClass to support different language engines"
