+++ soot-infoflow-android/src/soot/jimple/infoflow/android/source/ConfigurationBasedCategoryFilter.java
@@ -4,8 +4,7 @@ import soot.jimple.infoflow.InfoflowConfiguration.CategoryMode;
 import soot.jimple.infoflow.InfoflowConfiguration.SourceSinkFilterMode;
 import soot.jimple.infoflow.android.InfoflowAndroidConfiguration.SourceSinkConfiguration;
 import soot.jimple.infoflow.android.data.CategoryDefinition;
-import soot.jimple.infoflow.android.source.parsers.xml.AbstractXMLSourceSinkParser.ICategoryFilter;
-import soot.jimple.infoflow.sourcesSinks.definitions.ISourceSinkCategory;
+import soot.jimple.infoflow.android.source.parsers.xml.XMLSourceSinkParser.ICategoryFilter;
 import soot.jimple.infoflow.sourcesSinks.definitions.SourceSinkType;
 
 /**
@@ -29,29 +28,26 @@ public class ConfigurationBasedCategoryFilter implements ICategoryFilter {
 	}
 
 	@Override
-	public boolean acceptsCategory(ISourceSinkCategory category) {
+	public boolean acceptsCategory(CategoryDefinition category) {
 		// We cannot compare descriptions to the configuration file
-		if (category instanceof CategoryDefinition) {
-			CategoryDefinition catDef = (CategoryDefinition) category;
-			catDef = catDef.getIdOnlyDescription();
+		category = category.getIdOnlyDescription();
 
-			CategoryMode sourceType = config.getSourceCategoriesAndModes().get(catDef);
-			CategoryMode sinkType = config.getSinkCategoriesAndModes().get(catDef);
+		CategoryMode sourceType = config.getSourceCategoriesAndModes().get(category);
+		CategoryMode sinkType = config.getSinkCategoriesAndModes().get(category);
 
-			// Check whether the category is excluded for everything
-			if (sourceType != null && sourceType == CategoryMode.Exclude)
-				if (sinkType != null && sinkType == CategoryMode.Exclude)
-					return false;
-
-			// Check whether the category is included for something
-			if (config.getSinkFilterMode() == SourceSinkFilterMode.UseOnlyIncluded) {
-				if (sourceType != null && sourceType == CategoryMode.Include)
-					return true;
-				if (sinkType != null && sinkType == CategoryMode.Include)
-					return true;
-
+		// Check whether the category is excluded for everything
+		if (sourceType != null && sourceType == CategoryMode.Exclude)
+			if (sinkType != null && sinkType == CategoryMode.Exclude)
 				return false;
-			}
+
+		// Check whether the category is included for something
+		if (config.getSinkFilterMode() == SourceSinkFilterMode.UseOnlyIncluded) {
+			if (sourceType != null && sourceType == CategoryMode.Include)
+				return true;
+			if (sinkType != null && sinkType == CategoryMode.Include)
+				return true;
+
+			return false;
 		}
 
 		// There is no reason to exclude the category
@@ -59,36 +55,34 @@ public class ConfigurationBasedCategoryFilter implements ICategoryFilter {
 	}
 
 	@Override
-	public SourceSinkType filter(ISourceSinkCategory category, SourceSinkType sourceSinkType) {
+	public SourceSinkType filter(CategoryDefinition category, SourceSinkType sourceSinkType) {
 		// We cannot compare descriptions to the configuration file
-		if (category instanceof CategoryDefinition) {
-			CategoryDefinition catDef = (CategoryDefinition) category;
-			catDef = catDef.getIdOnlyDescription();
-			CategoryMode sourceMode = config.getSourceCategoriesAndModes().get(category);
-			CategoryMode sinkMode = config.getSinkCategoriesAndModes().get(category);
-			if (sourceSinkType == SourceSinkType.Source || sourceSinkType == SourceSinkType.Both) {
-				if (config.getSourceFilterMode() == SourceSinkFilterMode.UseAllButExcluded) {
-					if (sourceMode != null && sourceMode == CategoryMode.Exclude)
-						sourceSinkType = sourceSinkType.removeType(SourceSinkType.Source);
-				} else if (config.getSourceFilterMode() == SourceSinkFilterMode.UseOnlyIncluded) {
-					if (sourceMode == null || sourceMode != CategoryMode.Include)
-						sourceSinkType = sourceSinkType.removeType(SourceSinkType.Source);
-				}
-			}
+		category = category.getIdOnlyDescription();
 
-			if (sourceSinkType == SourceSinkType.Sink || sourceSinkType == SourceSinkType.Both) {
-				if (config.getSinkFilterMode() == SourceSinkFilterMode.UseAllButExcluded) {
-					if (sinkMode != null && sinkMode == CategoryMode.Exclude)
-						sourceSinkType = sourceSinkType.removeType(SourceSinkType.Sink);
-				} else if (config.getSinkFilterMode() == SourceSinkFilterMode.UseOnlyIncluded) {
-					if (sinkMode == null || sinkMode != CategoryMode.Include)
-						sourceSinkType = sourceSinkType.removeType(SourceSinkType.Sink);
-				}
+		CategoryMode sourceMode = config.getSourceCategoriesAndModes().get(category);
+		CategoryMode sinkMode = config.getSinkCategoriesAndModes().get(category);
+
+		if (sourceSinkType == SourceSinkType.Source || sourceSinkType == SourceSinkType.Both) {
+			if (config.getSourceFilterMode() == SourceSinkFilterMode.UseAllButExcluded) {
+				if (sourceMode != null && sourceMode == CategoryMode.Exclude)
+					sourceSinkType = sourceSinkType.removeType(SourceSinkType.Source);
+			} else if (config.getSourceFilterMode() == SourceSinkFilterMode.UseOnlyIncluded) {
+				if (sourceMode == null || sourceMode != CategoryMode.Include)
+					sourceSinkType = sourceSinkType.removeType(SourceSinkType.Source);
 			}
-			return sourceSinkType;
 		}
-		return SourceSinkType.Undefined;
 
+		if (sourceSinkType == SourceSinkType.Sink || sourceSinkType == SourceSinkType.Both) {
+			if (config.getSinkFilterMode() == SourceSinkFilterMode.UseAllButExcluded) {
+				if (sinkMode != null && sinkMode == CategoryMode.Exclude)
+					sourceSinkType = sourceSinkType.removeType(SourceSinkType.Sink);
+			} else if (config.getSinkFilterMode() == SourceSinkFilterMode.UseOnlyIncluded) {
+				if (sinkMode == null || sinkMode != CategoryMode.Include)
+					sourceSinkType = sourceSinkType.removeType(SourceSinkType.Sink);
+			}
+		}
+
+		return sourceSinkType;
 	}
 
 }
