+++ soot-infoflow/src/soot/jimple/infoflow/sourcesSinks/definitions/ISourceSinkDefinition.java
@@ -46,11 +46,4 @@ public interface ISourceSinkDefinition {
 	 */
 	public abstract void merge(ISourceSinkDefinition other);
 
-	/**
-	 * Indicates if the definition contains any sources or sinks
-	 * 
-	 * @return The boolean if this definition is empty
-	 */
-	public abstract boolean isEmpty();
-
 }
